Move entity selector to features
