+++ packages/datatrak-web/src/features/EntitySelector/SearchField.tsx
@@ -1,6 +1,6 @@
 /*
  * Tupaia
- *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ *  Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
  */
 
 import React from 'react';
@@ -8,7 +8,7 @@ import { TextField } from '@tupaia/ui-components';
 import styled from 'styled-components';
 import { Search, Clear } from '@material-ui/icons';
 import { InputAdornment, IconButton, TextFieldProps } from '@material-ui/core';
-import { InputHelperText } from '../../../components';
+import { InputHelperText } from '../../components';
 
 const StyledField = styled(TextField)<TextFieldProps>`
   margin-bottom: 0;
