+++ packages/datatrak-web/src/features/EntitySelector/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
+ */
+export { EntitySelector } from './EntitySelector';
