+++ packages/datatrak-web/src/features/Questions/EntityQuestion/EntityQuestion.tsx
@@ -3,45 +3,11 @@
  *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
 
-import React, { useState } from 'react';
-import styled from 'styled-components';
-import { useFormContext } from 'react-hook-form';
-import { FormHelperText, Typography } from '@material-ui/core';
-import { SpinningLoader, useDebounce } from '@tupaia/ui-components';
+import React from 'react';
 import { SurveyQuestionInputProps } from '../../../types';
-import { useProjectEntities, useEntityById } from '../../../api';
 import { useSurveyForm } from '../..';
-import { ResultsList } from './ResultsList';
-import { SearchField } from './SearchField';
+import { EntitySelector } from '../../EntitySelector';
 import { useEntityBaseFilters } from './utils';
-
-const Container = styled.div`
-  width: 100%;
-  z-index: 0;
-
-  fieldset:disabled & {
-    pointer-events: none;
-  }
-`;
-
-const Label = styled(Typography).attrs({
-  variant: 'h4',
-})`
-  font-size: 1rem;
-  cursor: pointer;
-`;
-
-const useSearchResults = (searchValue, config) => {
-  const filter = useEntityBaseFilters(config);
-  const { surveyProjectCode } = useSurveyForm();
-
-  const debouncedSearch = useDebounce(searchValue!, 200);
-  return useProjectEntities(surveyProjectCode, {
-    fields: ['id', 'parent_name', 'code', 'name', 'type'],
-    filter,
-    searchString: debouncedSearch,
-  });
-};
 
 export const EntityQuestion = ({
   id,
@@ -53,63 +19,28 @@ export const EntityQuestion = ({
   config,
 }: SurveyQuestionInputProps) => {
   const { isReviewScreen, isResponseScreen } = useSurveyForm();
-  const { errors } = useFormContext();
-  const [isDirty, setIsDirty] = useState(false);
-  const [searchValue, setSearchValue] = useState('');
 
-  // Display a previously selected value
-  useEntityById(value, {
-    staleTime: 0, // Needs to be 0 to make sure the entity is fetched on first render
-    enabled: !!value && !searchValue,
-    onSuccess: entityData => {
-      if (!isDirty) {
-        setSearchValue(entityData.name);
-      }
-    },
-  });
-  const onChangeSearch = newValue => {
-    setIsDirty(true);
-    setSearchValue(newValue);
-  };
+  const filter = useEntityBaseFilters(config);
 
-  const onSelect = entity => {
-    setIsDirty(true);
-    onChange(entity.value);
-  };
-
-  const { data: searchResults, isLoading, isFetched } = useSearchResults(searchValue, config);
-
-  const displayResults = searchResults?.filter(({ name: entityName }) => {
-    if (isDirty || !value) {
-      return true;
-    }
-    return entityName === searchValue;
-  });
+  const { surveyProjectCode } = useSurveyForm();
 
   return (
-    <Container>
-      {isReviewScreen || isResponseScreen ? (
-        <Label>{label}</Label>
-      ) : (
-        <SearchField
-          id={id}
-          isDirty={isDirty}
-          label={label}
-          detailLabel={detailLabel}
-          name={name!}
-          ref={ref}
-          onChangeSearch={onChangeSearch}
-          searchValue={searchValue}
-          invalid={invalid}
-          required={required}
-        />
-      )}
-      {errors && errors[name!] && <FormHelperText error>*{errors[name!].message}</FormHelperText>}
-      {!isFetched || isLoading ? (
-        <SpinningLoader />
-      ) : (
-        <ResultsList value={value} onSelect={onSelect} searchResults={displayResults} />
-      )}
-    </Container>
+    <EntitySelector
+      id={id}
+      label={label}
+      detailLabel={detailLabel}
+      name={name}
+      required={required}
+      controllerProps={{
+        onChange,
+        value,
+        ref,
+        invalid,
+      }}
+      config={config}
+      showLabel={isReviewScreen || isResponseScreen}
+      filter={filter}
+      projectCode={surveyProjectCode}
+    />
   );
 };
